feat(workflow): move the output resolvers into the catalog manifests
