+++ apps/web/src/components/workflow/nodes/core/list/schema.ts
@@ -3,18 +3,17 @@
 import { listManifest, type NodeManifest } from '@auxx/lib/workflow-engine/client'
 import type { NodeDefinition } from '~/components/workflow/types'
 import { defineFromManifest } from '../../define-from-manifest'
-import { computeListOutputVariables } from './output-variables'
 import type { ListNodeData } from './types'
 
 // The data half (operation vocabulary, config types, zod schemas, defaults,
-// validator, variable extraction) lives in the node catalog
+// validator, variable extraction, output resolver) lives in the node catalog
 // (`@auxx/lib/workflow-engine/catalog/nodes/list`). This file is the merge
-// site: manifest + the web-only output resolver.
+// site: manifest + the React parts.
 
 /** List node definition */
 export const listNodeDefinition: NodeDefinition<ListNodeData> = defineFromManifest(
   listManifest as unknown as NodeManifest<ListNodeData>,
-  { outputVariables: computeListOutputVariables }
+  {}
 )
 
 // Back-compat re-exports so no consumer import churns:
