+++ apps/web/src/components/workflow/nodes/core/resource-trigger/output-variables.ts
@@ -1,54 +1,3 @@
 // apps/web/src/components/workflow/nodes/core/resource-trigger/output-variables.ts
 
-import type { ResourceField } from '@auxx/lib/resources/client'
-import { generateResourceTriggerVariablesFromFields } from '@auxx/lib/workflow-engine/client'
-import type { OutputVariableContext } from '~/components/workflow/types/output-variables'
-import type { UnifiedVariable } from '~/components/workflow/types/variable-types'
-import type { ResourceTriggerData } from './types'
-
-/** Resource shape for variable generation (matches Find node) */
-type ResourceWithFields = {
-  id: string
-  label: string
-  plural: string
-  fields: ResourceField[]
-  entityDefinitionId?: string
-}
-
-/**
- * Get output variables for a resource trigger node
- * Unified function for both system resources and custom entities
- *
- * Pattern: Same as Find node - requires resource with fields for generation
- *
- * @param data - Resource trigger node data
- * @param nodeId - Node ID
- * @param context - Output variable context with resource access
- */
-export function getResourceTriggerOutputVariables(
-  data: ResourceTriggerData,
-  nodeId: string,
-  context: OutputVariableContext
-): UnifiedVariable[] {
-  const resource = context.resource as ResourceWithFields | undefined
-  const allResources = context.allResources as ResourceWithFields[]
-  // No resource selected yet - return empty (same as Find node)
-  if (!resource) {
-    return []
-  }
-
-  // Build resources map for relationship lookup
-  const resourcesMap = new Map(allResources?.map((r) => [r.id, r]) ?? [])
-
-  return generateResourceTriggerVariablesFromFields(
-    resource.fields,
-    {
-      id: resource.entityDefinitionId ?? resource.id,
-      label: resource.label,
-      plural: resource.plural,
-    },
-    nodeId,
-    data.operation,
-    { resourcesMap, maxDepth: 2 }
-  )
-}
+export { getResourceTriggerOutputVariables } from '@auxx/lib/workflow-engine/client'
